requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
distlib==0.3.7
Django==4.2.5
filelock==3.12.3
idna==3.4
mysql-connector-python==8.1.0
mysqlclient==2.2.0
numpy==1.26.0
opencv-contrib-python==4.8.1.78
platformdirs==3.10.0
protobuf==4.21.12
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.5
virtualenv==20.24.5
